FrameGrabber/requirements.txt
kafka-python>=2.0.0,<3.0.0
redis
PyYAML
opencv-python